requirements.txt
Flask==2.3.3
requests==2.31.0
Werkzeug==2.3.7
PyJWT==2.8.0
cryptography==41.0.4
pymysql==1.1.0
flask-cors==4.0.0
google-cloud-translate==2.0.1
python-dotenv==1.0.0